fix: PR comments and improvements for cookbook commands
